pyproject.toml
[project]
name = "recommender-systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.5.0",
    "hdbscan>=0.8.41",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "numpy==1.26.4",
    "optuna>=4.8.0",
    "pandas>=2.3.3",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "scikit-learn>=1.7.2",
    "scikit-surprise>=1.1.4",
    "seaborn>=0.13.2",
    "torch>=2.4.1",
]

# Specific versions of torch, torchvision, and torchaudio are installed based on the extra feature
[project.optional-dependencies]
cpu = ["torch==2.4.1", "torchvision==0.19.1", "torchaudio==2.4.1"]
cu124 = ["torch==2.4.1", "torchvision==0.19.1", "torchaudio==2.4.1"]

[tool.uv]
# The cpu and cu124 extras are mutually exclusive, you can't use both at the same time
conflicts = [[{ extra = "cpu" }, { extra = "cu124" }]]

[tool.uv.sources]
# The torch, torchvision, and torchaudio packages are installed from the pytorch-cpu and pytorch-cu124 indexes
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]
torchaudio = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]


[[tool.uv.index]]
# The pytorch-cpu index is used to install the torch, torchvision, and torchaudio packages
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
